requirements.txt
gspread==5.12.0
oauth2client==4.1.3
python-telegram-bot==21.4
python-dateutil==2.9.0
schedule==1.2.0